scalet2t

+++ mvn/datasets/utils.py
@@ -77,11 +77,11 @@ def cam2cam_batch(src, target, cameras_batch, batch_i):
 def cam2cam_precomputed_batch(src, target, cameras_batch, batch_i, cam2cams):
     if src == target:  # just apply intrinsics
         cam = cameras_batch[src][batch_i]  # multiview.Camera
-        return torch.cuda.FloatTensor(cam.intrinsics_padded)  # 3 x 4
+        return torch.cuda.DoubleTensor(cam.intrinsics_padded)  # 3 x 4
 
     K = cameras_batch[target][batch_i].intrinsics_padded
     return torch.mm(
-        torch.cuda.FloatTensor(K),
+        torch.cuda.DoubleTensor(K),
         cam2cams[batch_i][target]  # cam2cam from src -> target
     )  # 3 x 4
